Update to BETA 11.0.1 (5180)

+++ sources/org/telegram/tgnet/TLRPC$TL_themeSettings.java
@@ -58,7 +58,7 @@ public class TLRPC$TL_themeSettings extends TLRPC$ThemeSettings {
             int size = this.message_colors.size();
             abstractSerializedData.writeInt32(size);
             for (int i2 = 0; i2 < size; i2++) {
-                abstractSerializedData.writeInt32(this.message_colors.get(i2).intValue());
+                abstractSerializedData.writeInt32(((Integer) this.message_colors.get(i2)).intValue());
             }
         }
         if ((this.flags & 2) != 0) {
